refactor: move tests inside tests folder
