+++ src/normalizations/tests/normalize-soap-message-from-editor.test.ts
@@ -5,17 +5,17 @@
  * SPDX-License-Identifier: AGPL-3.0-only
  */
 
+import * as shell from '../../../__mocks__/@zextras/carbonio-shell-ui';
+import { generateEditor } from '../../commons/editor-generator';
+import { getIdentityItems } from '../../commons/get-identity-items';
+import { PARTICIPATION_STATUS } from '../../constants/api';
+import { ParticipationStatus } from '../../types/store/invite';
 import {
 	normalizeSoapMessageFromEditor,
 	setAlarmValue,
 	generateBodyRequest,
 	generateHtmlBodyRequest
-} from './normalize-soap-message-from-editor';
-import * as shell from '../../__mocks__/@zextras/carbonio-shell-ui';
-import { generateEditor } from '../commons/editor-generator';
-import { getIdentityItems } from '../commons/get-identity-items';
-import { PARTICIPATION_STATUS } from '../constants/api';
-import { ParticipationStatus } from '../types/store/invite';
+} from '../normalize-soap-message-from-editor';
 import { createFakeIdentity, getMockedAccountItem } from '@test-utils/accounts/fakeAccounts';
 
 const mainAccount = createFakeIdentity();
@@ -50,7 +50,7 @@ const sharedAccountEditorFolder = {
 
 const addressPrefKey = 'zimbraPrefFromAddress';
 
-jest.mock('../hooks/use-get-date-range-converted-to-timezone', () => ({
+jest.mock('../../hooks/use-get-date-range-converted-to-timezone', () => ({
 	getTimeStrings: jest.fn(() => 'Jan 1, 2024 10:00 AM - 11:00 AM')
 }));
